refactor(alloy): expose lightweight Tempo simulation env
