+++ .changelog/lightweight-alloy-revm.md
@@ -1,0 +1,5 @@
+---
+tempo-alloy: patch
+---
+
+Added a lightweight `revm` feature for Tempo transaction simulation without enabling the full Reth compatibility dependency set.
