+++ crates/alloy/src/rpc/mod.rs
@@ -9,6 +9,9 @@ pub use request::{FeeToken, TempoCallBuilderExt, TempoTransactionRequest};
 mod receipt;
 pub use receipt::TempoTransactionReceipt;
 
+#[cfg(feature = "revm")]
+mod revm_compat;
+
 #[cfg(feature = "reth")]
 mod reth_compat;
