+++ crates/alloy/src/rpc/reth_compat.rs
@@ -1,8 +1,7 @@
 use crate::rpc::{TempoHeaderResponse, TempoTransactionRequest};
 use alloy_consensus::{EthereumTxEnvelope, TxEip4844, error::ValueError};
 use alloy_network::{NetworkTransactionBuilder, TxSigner};
-use alloy_primitives::{Address, B256, Bytes, Signature};
-use core::num::NonZeroU64;
+use alloy_primitives::Signature;
 use reth_evm::EvmEnv;
 use reth_primitives_traits::SealedHeader;
 use reth_rpc_convert::{
@@ -11,18 +10,8 @@ use reth_rpc_convert::{
 use reth_rpc_eth_types::EthApiError;
 use tempo_chainspec::hardfork::TempoHardfork;
 use tempo_evm::TempoBlockEnv;
-use tempo_primitives::{
-    SignatureType, TempoHeader, TempoSignature, TempoTxEnvelope, TempoTxType,
-    transaction::{Call, RecoveredTempoAuthorization},
-};
-use tempo_revm::{TempoBatchCallEnv, TempoTxEnv};
-
-/// Non-zero transaction identifier used only for RPC simulations.
-///
-/// RPC requests are not final signed transactions, so gas filling and other request normalization
-/// can make a simulated signing payload differ from the eventual submitted transaction. Use a
-/// fixed sentinel instead of deriving a misleading future channel id from the simulated payload.
-const RPC_SIMULATION_UNIQUE_TX_IDENTIFIER: B256 = B256::new(*b"TEMPO_RPC_SIMULATION_MPP_CONTEXT");
+use tempo_primitives::{TempoHeader, TempoSignature, TempoTxEnvelope, TempoTxType};
+use tempo_revm::TempoTxEnv;
 
 impl TryIntoSimTx<TempoTxEnvelope> for TempoTransactionRequest {
     fn try_into_sim_tx(self) -> Result<TempoTxEnvelope, ValueError<Self>> {
@@ -105,207 +94,9 @@ impl TryIntoTxEnv<TempoTxEnv, TempoHardfork, TempoBlockEnv> for TempoTransaction
         self,
         evm_env: &EvmEnv<TempoHardfork, TempoBlockEnv>,
     ) -> Result<TempoTxEnv, Self::Err> {
-        let caller_addr = self.inner.from.unwrap_or_default();
-        let is_aa = self.has_aa_fields();
-
-        let fee_payer = if self.fee_payer_signature.is_some() {
-            // Try to recover the fee payer address from the signature.
-            // If recovery fails (e.g. dummy signature during gas estimation / fill),
-            // keep it unresolved so estimation/fill can continue with sender-paid semantics.
-            let recovered = self
-                .clone()
-                .build_aa()
-                .ok()
-                .and_then(|tx| tx.recover_fee_payer(caller_addr).ok());
-            Some(recovered)
-        } else {
-            None
-        };
-
-        let Self {
-            inner,
-            fee_token,
-            calls,
-            key_type,
-            key_data,
-            key_id,
-            tempo_authorization_list,
-            nonce_key,
-            key_authorization,
-            valid_before,
-            valid_after,
-            fee_payer_signature: _,
-        } = self;
-
-        Ok(TempoTxEnv {
-            fee_token,
-            is_system_tx: false,
-            unique_tx_identifier: Some(RPC_SIMULATION_UNIQUE_TX_IDENTIFIER),
-            fee_payer,
-            tempo_tx_env: if is_aa {
-                // Create mock signature for gas estimation
-                // If key_type is not provided, default to secp256k1
-                // For Keychain signatures, use the caller's address as the root key address
-                let key_type = key_type.unwrap_or(SignatureType::Secp256k1);
-                let mock_signature = create_mock_tempo_sig(
-                    &key_type,
-                    key_data.as_ref(),
-                    key_id,
-                    caller_addr,
-                    evm_env.spec_id().is_t1c(),
-                );
-
-                let mut calls = calls;
-                if let Some(to) = &inner.to {
-                    calls.push(Call {
-                        to: *to,
-                        value: inner.value.unwrap_or_default(),
-                        input: inner.input.clone().into_input().unwrap_or_default(),
-                    });
-                }
-                if calls.is_empty() {
-                    return Err(EthApiError::InvalidParams("empty calls list".to_string()));
-                }
-
-                Some(Box::new(TempoBatchCallEnv {
-                    aa_calls: calls,
-                    signature: mock_signature,
-                    tempo_authorization_list: tempo_authorization_list
-                        .into_iter()
-                        .map(RecoveredTempoAuthorization::new)
-                        .collect(),
-                    nonce_key: nonce_key.unwrap_or_default(),
-                    key_authorization,
-                    signature_hash: B256::ZERO,
-                    tx_hash: B256::ZERO,
-                    valid_before: valid_before.map(NonZeroU64::get),
-                    valid_after: valid_after.map(NonZeroU64::get),
-                    subblock_transaction: false,
-                    override_key_id: key_id,
-                    expiring_nonce_idx: None,
-                }))
-            } else {
-                None
-            },
-            inner: inner.try_into_tx_env(evm_env)?,
-        })
-    }
-}
-
-/// Creates a mock AA signature for gas estimation based on key type hints
-///
-/// - `key_type`: The primitive signature type (secp256k1, P256, WebAuthn)
-/// - `key_data`: Type-specific data (e.g., WebAuthn size)
-/// - `key_id`: If Some, wraps the signature in a Keychain wrapper (+3,000 gas for key validation)
-/// - `caller_addr`: The transaction caller address (used as root key address for Keychain)
-/// - `is_t1c`: Whether T1C is active — determines keychain signature version (V1 pre-T1C, V2 post-T1C)
-fn create_mock_tempo_sig(
-    key_type: &SignatureType,
-    key_data: Option<&Bytes>,
-    key_id: Option<Address>,
-    caller_addr: alloy_primitives::Address,
-    is_t1c: bool,
-) -> TempoSignature {
-    use tempo_primitives::transaction::tt_signature::{KeychainSignature, TempoSignature};
-
-    let inner_sig = create_mock_primitive_signature(key_type, key_data.cloned());
-
-    if key_id.is_some() {
-        // For Keychain signatures, the root_key_address is the caller (account owner).
-        let keychain_sig = if is_t1c {
-            KeychainSignature::new(caller_addr, inner_sig)
-        } else {
-            KeychainSignature::new_v1(caller_addr, inner_sig)
-        };
-        TempoSignature::Keychain(keychain_sig)
-    } else {
-        TempoSignature::Primitive(inner_sig)
-    }
-}
-
-/// Creates a mock primitive signature for gas estimation
-fn create_mock_primitive_signature(
-    sig_type: &SignatureType,
-    key_data: Option<Bytes>,
-) -> tempo_primitives::transaction::tt_signature::PrimitiveSignature {
-    use tempo_primitives::transaction::tt_signature::{
-        P256SignatureWithPreHash, PrimitiveSignature, WebAuthnSignature,
-    };
-
-    match sig_type {
-        SignatureType::Secp256k1 => {
-            // Create a dummy secp256k1 signature (65 bytes)
-            PrimitiveSignature::Secp256k1(Signature::new(
-                alloy_primitives::U256::ZERO,
-                alloy_primitives::U256::ZERO,
-                false,
-            ))
-        }
-        SignatureType::P256 => {
-            // Create a dummy P256 signature
-            PrimitiveSignature::P256(P256SignatureWithPreHash {
-                r: alloy_primitives::B256::ZERO,
-                s: alloy_primitives::B256::ZERO,
-                pub_key_x: alloy_primitives::B256::ZERO,
-                pub_key_y: alloy_primitives::B256::ZERO,
-                pre_hash: false,
-            })
-        }
-        SignatureType::WebAuthn => {
-            // Create a dummy WebAuthn signature with the specified size
-            // key_data contains the total size of webauthn_data (excluding 128 bytes for public keys)
-            // Default: 800 bytes if no key_data provided
-
-            // Base clientDataJSON template (50 bytes): {"type":"webauthn.get","challenge":"","origin":""}
-            // Authenticator data (37 bytes): 32 rpIdHash + 1 flags + 4 signCount
-            // Minimum total: 87 bytes
-            const BASE_CLIENT_JSON: &str = r#"{"type":"webauthn.get","challenge":"","origin":""}"#;
-            const AUTH_DATA_SIZE: usize = 37;
-            const MIN_WEBAUTHN_SIZE: usize = AUTH_DATA_SIZE + BASE_CLIENT_JSON.len(); // 87 bytes
-            const DEFAULT_WEBAUTHN_SIZE: usize = 800; // Default when no key_data provided
-            const MAX_WEBAUTHN_SIZE: usize = 8192; // Maximum realistic WebAuthn signature size
-
-            // Parse size from key_data, or use default
-            let size = if let Some(data) = key_data.as_ref() {
-                match data.len() {
-                    1 => data[0] as usize,
-                    2 => u16::from_be_bytes([data[0], data[1]]) as usize,
-                    4 => u32::from_be_bytes([data[0], data[1], data[2], data[3]]) as usize,
-                    _ => DEFAULT_WEBAUTHN_SIZE, // Fallback default
-                }
-            } else {
-                DEFAULT_WEBAUTHN_SIZE // Default size when no key_data provided
-            };
-
-            // Clamp size to safe bounds to prevent DoS via unbounded allocation
-            let size = size.clamp(MIN_WEBAUTHN_SIZE, MAX_WEBAUTHN_SIZE);
-
-            // Construct authenticatorData (37 bytes)
-            let mut webauthn_data = vec![0u8; AUTH_DATA_SIZE];
-            webauthn_data[32] = 0x01; // UP flag set
-
-            // Construct clientDataJSON with padding in origin field if needed
-            let additional_bytes = size - MIN_WEBAUTHN_SIZE;
-            let client_json = if additional_bytes > 0 {
-                // Add padding bytes to origin field
-                // {"type":"webauthn.get","challenge":"","origin":"XXXXX"}
-                let padding = "x".repeat(additional_bytes);
-                format!(r#"{{"type":"webauthn.get","challenge":"","origin":"{padding}"}}"#,)
-            } else {
-                BASE_CLIENT_JSON.to_string()
-            };
-
-            webauthn_data.extend_from_slice(client_json.as_bytes());
-            let webauthn_data = Bytes::from(webauthn_data);
-
-            PrimitiveSignature::WebAuthn(WebAuthnSignature {
-                webauthn_data,
-                r: alloy_primitives::B256::ZERO,
-                s: alloy_primitives::B256::ZERO,
-                pub_key_x: alloy_primitives::B256::ZERO,
-                pub_key_y: alloy_primitives::B256::ZERO,
-            })
-        }
+        let inner = self.inner.clone().try_into_tx_env(evm_env)?;
+        self.try_into_tempo_tx_env(TempoTxEnv::from(inner), evm_env.spec_id().is_t1c())
+            .map_err(|err| EthApiError::InvalidParams(err.to_string()))
     }
 }
 
@@ -338,13 +129,16 @@ impl FromConsensusHeader<TempoHeader> for TempoHeaderResponse {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use alloy_primitives::{TxKind, address};
+    use crate::rpc::revm_compat::{
+        RPC_SIMULATION_UNIQUE_TX_IDENTIFIER, create_mock_primitive_signature,
+    };
+    use alloy_primitives::{Address, B256, Bytes, TxKind, address};
     use alloy_rpc_types_eth::TransactionRequest;
     use alloy_signer::SignerSync;
     use alloy_signer_local::PrivateKeySigner;
     use reth_rpc_convert::TryIntoTxEnv;
     use tempo_primitives::{
-        TempoTransaction,
+        SignatureType, TempoTransaction,
         transaction::{Call, FEE_PAYER_SIGNATURE_MARKER, tt_signature::PrimitiveSignature},
     };
